feat: updated data change process and merged account settings JavaScript files
This commit refactors the data change process in the application to allow separate changes for username and name, as well as for changing the city by selecting a new city instead of a single data change for all fields.
Additionally, the JavaScript files for account settings have been merged into a single file for improved organization and maintenance.