feat: improved city matching in `get_city_weather` endpoint
Improved the city matching algorithm in the get_city_weather endpoint to ensure accurate weather data retrieval. Previously, there was a bug where the API sometimes returned information about the wrong city. This issue has been fixed by comparing the country of the city in the weather API response with the country of the city in the database. If they don't match, the API request is modified to include the city's country, resulting in correct weather data retrieval.